arm: dts: vf610-twr: Add support for sdhc1
The kernel was not able to boot from SD card because sdhc support
was not present into the dts.

A new entry for sdhc1 was added for vf610-twr board based on the
compatible entry present on imx53.

After applying these changes, the kernel is able to boot successfully
from SD card.

Signed-off-by: Cosmin Stoica <cosminstefan.stoica@freescale.com>